Fixed issue on blog post page
