+++ src/Pages/Admission/Tuition And Scholarship/TuitionAndScholarship.jsx
@@ -17,7 +17,6 @@ import {
   AccordionItem,
   AccordionButton,
   AccordionPanel,
-  Link,
 } from "@chakra-ui/react";
 import { FaChevronCircleDown } from "react-icons/fa";
 import { FaChevronCircleUp } from "react-icons/fa";
@@ -33,13 +32,13 @@ const TuitionAndScholarship = () => {
       year1: "YEAR 1",
       year2: "YEAR 2",
       tuitionFee: "TUITION FEE",
-      tuitionFeePrice1: "#369,000",
-      tuitionFeePrice2: "#405,000",
+      tuitionFeePrice1: "#502,800",
+      tuitionFeePrice2: "#526,800",
       acceptanceFee: "ACCEPTANCE FEE",
       acceptanceFeePrice1: "#5,000",
       acceptanceFeePrice2: "--",
       examinationFee: "ADMIN FEE",
-      examinationFeePrice1: "#42,000",
+      examinationFeePrice1: "#2,000",
       examinationFeePrice2: "#40,000",
       workshopFee: "PRACTICAL UNIFORMS",
       workshopFeePrice1: "#35,000",
@@ -49,25 +48,25 @@ const TuitionAndScholarship = () => {
       devFeePrice2: "#40,000",
       practicalFee: "MANUALS/WORKBOOKS",
       practicalFeePrice1: "#7,500",
-      practicalFeePrice2: "#10,000",
+      practicalFeePrice2: "#7,500",
       studentID: "MEDICALS",
-      studentIDPrice1: "#15,000",
+      studentIDPrice1: "#19,000",
       studentIDPrice2: "--",
       studentPracticals: "STUDENT PRACTICALS*",
-      studentPracticalsPrice1: "#60,000",
-      studentPracticalsPrice2: "#60,000",
+      studentPracticalsPrice1: "#60,500",
+      studentPracticalsPrice2: "#60,500",
       hmoScheme: "HMO SCHEME",
-      hmoSchemePrice1: "#9,000",
-      hmoSchemePrice2: "#9,000",
+      hmoSchemePrice1: "#10,000",
+      hmoSchemePrice2: "#10,000",
       fees: "FEES TO BE PAID",
-      feesPrice1: "#94,000",
-      feesPrice2: "#110,000",
+      feesPrice1: "#144,000",
+      feesPrice2: "#144,000",
       discount: "AMOUNT SUBSIDIZED BY COLLEGE",
-      discountPrice1: "#275,000",
-      discountPrice2: "#295,000",
+      discountPrice1: "#358,800",
+      discountPrice2: "#382,800",
       totalFees: "TOTAL FEES TO BE PAID",
-      totalFeesPrice1: "#307,500",
-      totalFeesPrice2: "#269,000",
+      totalFeesPrice1: "#299,000",
+      totalFeesPrice2: "#262,000",
     },
 
     {
@@ -76,13 +75,13 @@ const TuitionAndScholarship = () => {
       year1: "YEAR 1",
       year2: "YEAR 2",
       tuitionFee: "TUITION FEE",
-      tuitionFeePrice1: "#377,000",
-      tuitionFeePrice2: "#374,000",
+      tuitionFeePrice1: "#496,800",
+      tuitionFeePrice2: "#490,800",
       acceptanceFee: "ACCEPTANCE FEE",
       acceptanceFeePrice1: "#5,000",
       acceptanceFeePrice2: "--",
       examinationFee: "ADMIN FEE",
-      examinationFeePrice1: "#41,000",
+      examinationFeePrice1: "#9,500",
       examinationFeePrice2: "#40,000",
       workshopFee: "PRACTICAL UNIFORMS",
       workshopFeePrice1: "#35,000",
@@ -92,25 +91,25 @@ const TuitionAndScholarship = () => {
       devFeePrice2: "#40,000",
       practicalFee: "MANUALS/WORKBOOKS",
       practicalFeePrice1: "#12,500",
-      practicalFeePrice2: "#15,000",
+      practicalFeePrice2: "#12,500",
       studentID: "MEDICALS",
       studentIDPrice1: "#28,500",
       studentIDPrice2: "--",
       studentPracticals: "STUDENT PRACTICALS*",
-      studentPracticalsPrice1: "#60,000",
-      studentPracticalsPrice2: "#60,000",
+      studentPracticalsPrice1: "#60,500",
+      studentPracticalsPrice2: "#60,500",
       hmoScheme: "HMO SCHEME",
-      hmoSchemePrice1: "#9,000",
-      hmoSchemePrice2: "#9,000",
+      hmoSchemePrice1: "#10,000",
+      hmoSchemePrice2: "#10,000",
       fees: "FEES TO BE PAID",
-      feesPrice1: "#91,000",
-      feesPrice2: "#99,000",
+      feesPrice1: "#128,000",
+      feesPrice2: "#139,000",
       discount: "AMOUNT SUBSIDIZED BY COLLEGE",
-      discountPrice1: "#286,000",
-      discountPrice2: "#275,000",
+      discountPrice1: "#368,800",
+      discountPrice2: "#351,800",
       totalFees: "TOTAL FEES TO BE PAID",
-      totalFeesPrice1: "#322,000",
-      totalFeesPrice2: "#263,000",
+      totalFeesPrice1: "#295,000",
+      totalFeesPrice2: "#262,000",
     },
 
     {
@@ -118,15 +117,15 @@ const TuitionAndScholarship = () => {
       item: "ITEM",
       year1: "YEAR 1",
       tuitionFee: "TUITION FEE",
-      tuitionFeePrice1: "#376,000",
+      tuitionFeePrice1: "#472,800",
       discount: "AMOUNT SUBSIDIZED BY COLLEGE",
-      discountPrice1: "#278,000",
+      discountPrice1: "#356,800",
       fees: "FEES TO BE PAID",
-      feesPrice1: "#98,000",
+      feesPrice1: "#116,000",
       acceptanceFee: "ACCEPTANCE FEE",
       acceptanceFeePrice1: "#5,000",
       examinationFee: "ADMIN FEE",
-      examinationFeePrice1: "#29,000",
+      examinationFeePrice1: "#16,000",
       workshopFee: "PRACTICAL UNIFORMS",
       workshopFeePrice1: "#35,000",
       devFee: "DEVELOPMENT FEE",
@@ -136,17 +135,17 @@ const TuitionAndScholarship = () => {
       studentID: "MEDICALS",
       studentIDPrice1: "#28,500",
       studentPracticals: "STUDENT PRACTICALS*",
-      studentPracticalsPrice1: "#70,000",
+      studentPracticalsPrice1: "#70,500",
       hmoScheme: "HMO SCHEME",
-      hmoSchemePrice1: "#9,000",
+      hmoSchemePrice1: "#10,000",
 
       totalFees: "TOTAL FEES TO BE PAID",
-      totalFeesPrice1: "#304,000",
+      totalFeesPrice1: "#277,500",
     },
   ];
   const navigate = useNavigate();
   const redirect = () => {
-    navigate("", { replace: true });
+    navigate("/support", { replace: true });
   };
 
   return (
@@ -662,25 +661,25 @@ const TuitionAndScholarship = () => {
                       You can Support the Scholarship Scheme by following the
                       link below:
                     </Heading>
-                    <Link
-                      href="/support"
+                    <Text
+                      onClick={() => redirect()}
                       bg="#021D37"
                       p="10px"
                       w="142px"
-                      height="46.89px"
                       color="#FFF"
                       borderRadius="5px"
                       fontFamily="Open Sans"
                       fontWeight="700"
                       fontSize="17px"
                       textAlign="center"
+                      cursor="pointer"
                       _hover={{
                         textDecoration: "none",
                         bg: "#020E1B",
                       }}
                     >
                       SUPPORT US
-                    </Link>
+                    </Text>
                   </Box>
                 </AccordionPanel>
               </Box>
